Adds the colored rivers to the Berlin Map

Adds the colored rivers to the Berlin Map

add_coloredRivers(
  ext_riversList,
  sixBreaks,
  dataType = "time",
  LegendTitle,
  LegendLocation
)

Arguments

ext_riversList

The list of Berlin rivers loaded with load_berlin_rivers(), where every table is extended with extend_riverTable()

sixBreaks

Breaks defining the lower limits of the categories. Must be the same that were used in extend_riverTable()

dataType

Is needed for the Legend. If "time" is used, it is assumed that the first water quality category is between 0 and a value above 0, while for the counting of events the first category is 0 only. If that is wanted anything but "time" can be assigned.

LegendTitle

String with title of the legend

LegendLocation

Either "top" or "right" (outside the plot margin)