Creates and stores a map Plot of all oxygen deficiencies Events

Creates and stores a map Plot of all oxygen deficiencies Events

mapPlot_EventsNumber(
  BerlinRivers,
  df_aggr,
  scenarioName = "",
  savingPath = "",
  varName = "events",
  sixBreaks = c(-1, 0, 1, 3, 6, 10),
  decoupling = "",
  dec = ","
)

Arguments

BerlinRivers

Berlin rivers loaded from package with load_berlin_rivers()

df_aggr

Aggregated misa assessment data frame

scenarioName

A String, only used for the plot title and file name

savingPath

Directory where plot is going to be stored. If left empty (default) a new device is opened in RStudio for the plot.

varName

Column name that is used for water quality categorisation

sixBreaks

6 low limits of water quality categories

decoupling

Character vector defining a decoupling scenario. Default is "" for no scenario. Possible scenarios can be obtained by the column names

dec

What should be used a decimal character? Default is "," since the plots are in German language